TFP-5702 flytter dagens opprydding av overstyring av dekningsgrad til riktig steg
